smartmenu/requirements.txt
fastapi
uvicorn
openai
python-dotenv
PyPDF2
pytesseract
pdf2image
Pillow
qrcode
python-multipart
